Message cleanup
git-svn-id: file:///home/jj/hercules.svn/trunk@6973 956126f8-22a0-4046-8f4a-272fa8102e63

+++ comm3705.c
@@ -975,7 +975,7 @@ static void logdump(char *txt,DEVBLK *dev,BYTE *bfr,size_t sz)
         return;
     }
     WRMSG(HHC01048,"D",SSID_TO_LCSS(dev->ssid),dev->devnum,txt);
-    WRMSG(HHC01049,"D",SSID_TO_LCSS(dev->ssid),dev->devnum,txt,sz,sz);
+    WRMSG(HHC01049,"D",SSID_TO_LCSS(dev->ssid),dev->devnum,txt,(u_int)sz,(u_int)sz);
     buf[0] = 0;
     for(i=0;i<sz;i++)
     {
